// src/modules/ecosim/assets/types.d.ts
// Generated by https://quicktype.io

export type AssetResponse = {
    IsValid: boolean;
    Data: Data;
};

// Generated by https://quicktype.io

export type AssetInfo = {
    id: number;
    assetType: number;
    name: string;
    description: null | string;
    genres: string[];
    creatorType: string;
    creatorTargetId: number;
    creatorName: string;
    offsaleDeadline: null | string;
    itemRestrictions: any[];
    saleCount: number;
    itemType: string;
    favoriteCount: number;
    isForSale: boolean;
    price: number | null;
    priceTickets: number | null;
    lowestPrice: null | number;
    priceStatus: null | string;
    lowestSellerData?: any | null;
    unitsAvailableForConsumption?: any | null;
    serialCount: number;
    is18Plus: boolean;
    moderationStatus: string;
    createdAt: string;
    updatedAt: string;
};

export type Data = {
    TotalItems: number;
    Start: number;
    End: number;
    Page: number;
    nextPageCursor: string | null;
    previousPageCursor: string | null;
    ItemsPerPage: number;
    PageType: string;
    Items: ItemElement[];
};

export type ItemElement = {
    AssetRestrictionIcon: AssetRestrictionIcon;
    Item: ItemItem;
    Creator: Creator;
    Product: Product;
    PrivateSeller: null;
    Thumbnail: {};
    UserItem: {};
};

export type AssetRestrictionIcon = {
    CssTag: CSSTag;
};

export enum CSSTag {
    Empty = "",
    Limited = "limited",
    LimitedUnique = "limited-unique",
}

export type Creator = {
    Id: number;
    Name: string;
    Type: number;
    CreatorProfileLink: string;
};

export type ItemItem = {
    AssetId: number;
    UniverseId: number | null;
    Name: string;
    AbsoluteUrl: string;
    AssetType: number;
};

export type Product = {
    PriceInRobux: number;
    SerialNumber: number | null;
};

export type AssetSerach = {
    _total: number;
    keyword: string;
    elasticsearchDebugInfo?: null;
    nextPageCursor: null | string;
    previousPageCursor: null | string;
    data: AssetSmall[];
};

export type AssetSmall = {
    itemType: "Asset";
    id: number;
};